Move AutoBatcher to its own separate file.
